requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
pandas
scikit-learn
joblib
mlflow
PyYAML
dvc
